feat: coder routine

+++ includes/time.h
@@ -2,6 +2,6 @@
 # define TIME_H
 
 void	start_timestamp(void);
-int		get_timestamp(void);
+long	get_timestamp(void);
 
 #endif
